Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,8 +2,8 @@
 
 int main() {
     float rad;
-    scanf("%.1f", &rad);
-    int area = 3.14*rad*rad;
+    scanf("%f", &rad);
+    float area = 3.14*rad*rad;
     printf("Area: %.2f",area)
    
     return 0;
